requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.7.0
django-filer==2.0.2
django-formtools==2.3
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-text-ckeditor==4.0.0
easy-thumbnails==2.7.1
html5lib==1.1
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
Unidecode==1.1.2
urllib3==1.26.5
webencodings==0.5.1
wincertstore==0.2